Handle conditionalDependencies
I propose to add `conditionalDependencies` (`optionalDependencies` name is
already taken by npm and doesn't have the needed mechanics behind it).

The suggestion is to use the same semantics as for `peerDependencies` but not
fail if those packages mentioned in conditionalDependencies are absent from
installation.

If package `optdep` is listed in `conditionalDependencies` for the package `pkg`:

* `optdep` is not being installed automatically (same as with
  `peerDependencies`) for the `pkg`
* Unlike with peerDependencies it is not an error to have no `optdep` installed
  at all.
* If `optdep` is present in the installation (required by some of packages
  above) then it's being pulled into `pkg`'s env.

Fixes #46